refactor: decomposing nav into separate files
